fixed (many) errors

+++ libft/libft.h
@@ -34,7 +34,7 @@ char			*ft_strchr(const char *s, int c);
 char			*ft_strdup(const char *s1);
 size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
 size_t			ft_strlcpy(char *dst, const char *src, size_t destsize);
-size_t			ft_strlen(const char *s);
+int				ft_strlen(const char *s);
 int				ft_strncmp(const char *s1, const char *s2, size_t n);
 char			*ft_strnstr(const char *haystack, const char *needle,
 					size_t len);
